# implement generic count-min-log sketch with zero allocations