--- a/go.mod
+++ b/go.mod
@@ -1,0 +1,5 @@
+module github.com/seiflotfy/count-min-log
+
+go 1.23.4
+
+require github.com/dgryski/go-farm v0.0.0-20240924180020-3414d57e47da