Rewrite CML sketch to use generics for flexible register sizes (uint8, uint16, uint32). Optimize data layout to flat array for better cache locality. Eliminate all allocations during Update/BulkUpdate operations, reducing memory overhead by 100%.

Improve mathematical operations:
- Replace manual exp calculation with precomputed logExp value
- Use math.Exp for faster and more accurate calculations
- Optimize probability calculations in increaseDecision function
- Use proper PCG random number generator with deterministic seeding

Add complete implementation with:
- Marshaling/unmarshaling support for persistence
- Sketch merging for distributed processing
- Robust test suite with accuracy verification
- Hash value reuse across operations

Performance improves dramatically across all operations:
- Update: ~67% faster
- BulkUpdate: ~74% faster
- Query: ~60% faster

```
goos: darwin
goarch: arm64
pkg: github.com/seiflotfy/count-min-log
cpu: Apple M3 Max
                                                           │    before    │               after                │
                                                           │    sec/op    │   sec/op     vs base               │
Sketch/fn=Update/cap=1000000/err=0.010-16                     60.13n ± 1%   20.11n ± 1%  -66.56% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=1000000/err=0.010/freq=10-16        324.40n ± 3%   92.43n ± 1%  -71.51% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=1000000/err=0.010/freq=100-16       3051.5n ± 2%   802.2n ± 2%  -73.71% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=1000000/err=0.010/freq=1000-16      30.792µ ± 2%   7.929µ ± 1%  -74.25% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=1000000/err=0.010/freq=10000-16     308.99µ ± 0%   79.79µ ± 1%  -74.18% (p=0.002 n=6)
Sketch/fn=Query/cap=1000000/err=0.010-16                      37.34n ± 1%   14.85n ± 1%  -60.23% (p=0.002 n=6)
Sketch/fn=Update/cap=10000000/err=0.010-16                    62.46n ± 4%   20.19n ± 1%  -67.68% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=10000000/err=0.010/freq=10-16       331.25n ± 1%   92.79n ± 2%  -71.99% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=10000000/err=0.010/freq=100-16      3065.0n ± 0%   808.5n ± 1%  -73.62% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=10000000/err=0.010/freq=1000-16     31.036µ ± 2%   7.960µ ± 1%  -74.35% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=10000000/err=0.010/freq=10000-16    316.50µ ± 0%   79.50µ ± 1%  -74.88% (p=0.002 n=6)
Sketch/fn=Query/cap=10000000/err=0.010-16                     38.23n ± 1%   14.90n ± 1%  -61.02% (p=0.002 n=6)
Sketch/fn=Update/cap=100000000/err=0.010-16                   68.50n ± 4%   20.04n ± 0%  -70.74% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=100000000/err=0.010/freq=10-16      345.20n ± 0%   91.80n ± 1%  -73.41% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=100000000/err=0.010/freq=100-16     3196.5n ± 1%   809.9n ± 1%  -74.66% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=100000000/err=0.010/freq=1000-16    31.573µ ± 0%   7.966µ ± 1%  -74.77% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=100000000/err=0.010/freq=10000-16   315.84µ ± 0%   79.42µ ± 1%  -74.85% (p=0.002 n=6)
Sketch/fn=Query/cap=100000000/err=0.010-16                    38.32n ± 1%   14.96n ± 1%  -60.97% (p=0.002 n=6)
geomean                                                       1.703µ        491.7n       -71.13%

                                                           │    before    │                 after                  │
                                                           │     B/op     │    B/op     vs base                    │
Sketch/fn=Update/cap=1000000/err=0.010-16                    64.00 ± 0%      0.00 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=1000000/err=0.010/freq=10-16        64.00 ± 0%      0.00 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=1000000/err=0.010/freq=100-16       64.00 ± 0%      0.00 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=1000000/err=0.010/freq=1000-16      64.00 ± 0%      0.00 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=1000000/err=0.010/freq=10000-16     64.00 ± 0%      0.00 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=Query/cap=1000000/err=0.010-16                     0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=6) ¹
Sketch/fn=Update/cap=10000000/err=0.010-16                   64.00 ± 0%      0.00 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=10000000/err=0.010/freq=10-16       64.00 ± 0%      0.00 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=10000000/err=0.010/freq=100-16      64.00 ± 0%      0.00 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=10000000/err=0.010/freq=1000-16     64.00 ± 0%      0.00 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=10000000/err=0.010/freq=10000-16    64.00 ± 0%      0.00 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=Query/cap=10000000/err=0.010-16                    0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=6) ¹
Sketch/fn=Update/cap=100000000/err=0.010-16                  64.00 ± 0%      0.00 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=100000000/err=0.010/freq=10-16      64.00 ± 0%      0.00 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=100000000/err=0.010/freq=100-16     64.00 ± 0%      0.00 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=100000000/err=0.010/freq=1000-16    64.00 ± 0%      0.00 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=100000000/err=0.010/freq=10000-16   64.00 ± 0%      0.00 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=Query/cap=100000000/err=0.010-16                   0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=6) ¹
geomean                                                                 ²               ?                      ² ³
¹ all samples are equal
² summaries must be >0 to compute geomean
³ ratios must be >0 to compute geomean

                                                           │    before    │                 after                  │
                                                           │  allocs/op   │ allocs/op   vs base                    │
Sketch/fn=Update/cap=1000000/err=0.010-16                    1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=1000000/err=0.010/freq=10-16        1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=1000000/err=0.010/freq=100-16       1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=1000000/err=0.010/freq=1000-16      1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=1000000/err=0.010/freq=10000-16     1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=Query/cap=1000000/err=0.010-16                     0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=6) ¹
Sketch/fn=Update/cap=10000000/err=0.010-16                   1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=10000000/err=0.010/freq=10-16       1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=10000000/err=0.010/freq=100-16      1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=10000000/err=0.010/freq=1000-16     1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=10000000/err=0.010/freq=10000-16    1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=Query/cap=10000000/err=0.010-16                    0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=6) ¹
Sketch/fn=Update/cap=100000000/err=0.010-16                  1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=100000000/err=0.010/freq=10-16      1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=100000000/err=0.010/freq=100-16     1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=100000000/err=0.010/freq=1000-16    1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=BulkUpdate/cap=100000000/err=0.010/freq=10000-16   1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.002 n=6)
Sketch/fn=Query/cap=100000000/err=0.010-16                   0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=6) ¹
geomean                                                                 ²               ?                      ² ³
¹ all samples are equal
² summaries must be >0 to compute geomean
³ ratios must be >0 to compute geomean
```